requirement.txt
telebot>=0.0.5
asyncio>=3.4.3
python-dotenv>=1.0.1
aiohttp>=3.9.3